% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sample3_function.R
\name{sample3_function}
\alias{sample3_function}
\title{An example function}
\usage{
sample3_function(input = TRUE)
}
\arguments{
\item{input}{Is it TRUE/FALSE}
}
\description{
This function checks for the input argument being passed
}
\examples{
sample3_function()
}
\keyword{sample3}

